add an optional prop to set min-width for button
