+++ components/button/src/overrides/button.tsx
@@ -32,7 +32,10 @@ BlenderButton.displayName = 'BlenderButton';
 // Styles
 // ==============================
 
-const buttonStyles = (_: any, { tag, type, look, size, soft, block, justify, disabled }: any) => {
+const buttonStyles = (
+	_: any,
+	{ tag, type, look, size, soft, block, justify, disabled, horizontalPadding, minWidth }: any
+) => {
 	const mq = useMediaQuery();
 	const { COLORS, TYPE } = useBrand();
 
@@ -191,6 +194,7 @@ const buttonStyles = (_: any, { tag, type, look, size, soft, block, justify, dis
 				padding: 0,
 			},
 		}),
+
 		// =========
 
 		label: 'button',
@@ -209,6 +213,7 @@ const buttonStyles = (_: any, { tag, type, look, size, soft, block, justify, dis
 		boxSizing: 'border-box',
 		display: blockArr.map((b) => b !== null && (b ? 'flex' : 'inline-flex')),
 		width: blockArr.map((b) => b !== null && (b ? '100%' : 'auto')),
+		minWidth: minWidth,
 		...(look !== 'unstyled' && {
 			fontSize: sizeArr.map((s) => s && sizeMap[s].fontSize),
 			...TYPE.bodyFont[400],
@@ -218,6 +223,13 @@ const buttonStyles = (_: any, { tag, type, look, size, soft, block, justify, dis
 			borderRadius: look !== 'link' && '0.1875rem',
 			transition: 'background 0.2s ease, color 0.2s ease',
 		}),
+
+		// Trim horizontal padding if button is link style and false for horizontalPadding prop
+		...(!horizontalPadding &&
+			look === 'link' && {
+				paddingLeft: 0,
+				paddingRight: 0,
+			}),
 		...styleMap[look][soft ? 'softCSS' : 'standardCSS'],
 
 		// Hover state (but excluded if disabled or inside a disabled fieldset)
